Refactored most of the Game logic into its own package. Added some comments
